backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
motor==3.1.1
pydantic==2.8.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==2.1.1
